requirements.txt
fastapi
uvicorn
tensorflow
keras
numpy
pandas
scikit-learn
Pillow
faiss-cpu